+++ actix-tls/src/accept/openssl.rs
@@ -1,3 +1,5 @@
+//! OpenSSL based acceptor service.
+
 use std::{
     convert::Infallible,
     future::Future,
@@ -16,7 +18,6 @@ use actix_rt::{
 use actix_service::{Service, ServiceFactory};
 use actix_utils::counter::{Counter, CounterGuard};
 use futures_core::future::LocalBoxFuture;
-
 pub use openssl::ssl::{
     AlpnError, Error as SslError, HandshakeError, Ssl, SslAcceptor, SslAcceptorBuilder,
 };
@@ -24,30 +25,30 @@ use pin_project_lite::pin_project;
 
 use super::{TlsError, DEFAULT_TLS_HANDSHAKE_TIMEOUT, MAX_CONN_COUNTER};
 
-/// Wrapper type for `tokio_openssl::SslStream` in order to impl `ActixStream` trait.
-pub struct TlsStream<T>(tokio_openssl::SslStream<T>);
+/// Wraps a [`tokio_openssl::SslStream`] in order to impl [`ActixStream`] trait.
+pub struct TlsStream<IO>(tokio_openssl::SslStream<IO>);
 
-impl<T> From<tokio_openssl::SslStream<T>> for TlsStream<T> {
-    fn from(stream: tokio_openssl::SslStream<T>) -> Self {
+impl<IO> From<tokio_openssl::SslStream<IO>> for TlsStream<IO> {
+    fn from(stream: tokio_openssl::SslStream<IO>) -> Self {
         Self(stream)
     }
 }
 
-impl<T> Deref for TlsStream<T> {
-    type Target = tokio_openssl::SslStream<T>;
+impl<IO> Deref for TlsStream<IO> {
+    type Target = tokio_openssl::SslStream<IO>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl<T> DerefMut for TlsStream<T> {
+impl<IO> DerefMut for TlsStream<IO> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
 }
 
-impl<T: ActixStream> AsyncRead for TlsStream<T> {
+impl<IO: ActixStream> AsyncRead for TlsStream<IO> {
     fn poll_read(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -57,7 +58,7 @@ impl<T: ActixStream> AsyncRead for TlsStream<T> {
     }
 }
 
-impl<T: ActixStream> AsyncWrite for TlsStream<T> {
+impl<IO: ActixStream> AsyncWrite for TlsStream<IO> {
     fn poll_write(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -87,13 +88,13 @@ impl<T: ActixStream> AsyncWrite for TlsStream<T> {
     }
 }
 
-impl<T: ActixStream> ActixStream for TlsStream<T> {
+impl<IO: ActixStream> ActixStream for TlsStream<IO> {
     fn poll_read_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
-        T::poll_read_ready((&**self).get_ref(), cx)
+        IO::poll_read_ready((&**self).get_ref(), cx)
     }
 
     fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
-        T::poll_write_ready((&**self).get_ref(), cx)
+        IO::poll_write_ready((&**self).get_ref(), cx)
     }
 }
 
@@ -134,8 +135,8 @@ impl Clone for Acceptor {
     }
 }
 
-impl<T: ActixStream> ServiceFactory<T> for Acceptor {
-    type Response = TlsStream<T>;
+impl<IO: ActixStream> ServiceFactory<IO> for Acceptor {
+    type Response = TlsStream<IO>;
     type Error = TlsError<SslError, Infallible>;
     type Config = ();
     type Service = AcceptorService;
@@ -155,16 +156,17 @@ impl<T: ActixStream> ServiceFactory<T> for Acceptor {
     }
 }
 
+/// OpenSSL based acceptor service.
 pub struct AcceptorService {
     acceptor: SslAcceptor,
     conns: Counter,
     handshake_timeout: Duration,
 }
 
-impl<T: ActixStream> Service<T> for AcceptorService {
-    type Response = TlsStream<T>;
+impl<IO: ActixStream> Service<IO> for AcceptorService {
+    type Response = TlsStream<IO>;
     type Error = TlsError<SslError, Infallible>;
-    type Future = AcceptorServiceResponse<T>;
+    type Future = AcceptFut<IO>;
 
     fn poll_ready(&self, ctx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         if self.conns.available(ctx) {
@@ -174,11 +176,11 @@ impl<T: ActixStream> Service<T> for AcceptorService {
         }
     }
 
-    fn call(&self, io: T) -> Self::Future {
+    fn call(&self, io: IO) -> Self::Future {
         let ssl_ctx = self.acceptor.context();
         let ssl = Ssl::new(ssl_ctx).expect("Provided SSL acceptor was invalid.");
 
-        AcceptorServiceResponse {
+        AcceptFut {
             _guard: self.conns.get(),
             timeout: sleep(self.handshake_timeout),
             stream: Some(tokio_openssl::SslStream::new(ssl, io).unwrap()),
@@ -187,16 +189,18 @@ impl<T: ActixStream> Service<T> for AcceptorService {
 }
 
 pin_project! {
-    pub struct AcceptorServiceResponse<T: ActixStream> {
-        stream: Option<tokio_openssl::SslStream<T>>,
+    /// Accept future for Rustls service.
+    #[doc(hidden)]
+    pub struct AcceptFut<IO: ActixStream> {
+        stream: Option<tokio_openssl::SslStream<IO>>,
         #[pin]
         timeout: Sleep,
         _guard: CounterGuard,
     }
 }
 
-impl<T: ActixStream> Future for AcceptorServiceResponse<T> {
-    type Output = Result<TlsStream<T>, TlsError<SslError, Infallible>>;
+impl<IO: ActixStream> Future for AcceptFut<IO> {
+    type Output = Result<TlsStream<IO>, TlsError<SslError, Infallible>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let this = self.project();