+++ actix-tls/src/accept/rustls.rs
@@ -1,3 +1,5 @@
+//! Rustls based acceptor service.
+
 use std::{
     convert::Infallible,
     future::Future,
@@ -18,36 +20,35 @@ use actix_service::{Service, ServiceFactory};
 use actix_utils::counter::{Counter, CounterGuard};
 use futures_core::future::LocalBoxFuture;
 use pin_project_lite::pin_project;
-use tokio_rustls::{Accept, TlsAcceptor};
-
 pub use tokio_rustls::rustls::ServerConfig;
+use tokio_rustls::{Accept, TlsAcceptor};
 
 use super::{TlsError, DEFAULT_TLS_HANDSHAKE_TIMEOUT, MAX_CONN_COUNTER};
 
-/// Wrapper type for `tokio_openssl::SslStream` in order to impl `ActixStream` trait.
-pub struct TlsStream<T>(tokio_rustls::server::TlsStream<T>);
+/// Wraps a [`tokio_rustls::server::TlsStream`] in order to impl [`ActixStream`] trait.
+pub struct TlsStream<IO>(tokio_rustls::server::TlsStream<IO>);
 
-impl<T> From<tokio_rustls::server::TlsStream<T>> for TlsStream<T> {
-    fn from(stream: tokio_rustls::server::TlsStream<T>) -> Self {
+impl<IO> From<tokio_rustls::server::TlsStream<IO>> for TlsStream<IO> {
+    fn from(stream: tokio_rustls::server::TlsStream<IO>) -> Self {
         Self(stream)
     }
 }
 
-impl<T> Deref for TlsStream<T> {
-    type Target = tokio_rustls::server::TlsStream<T>;
+impl<IO> Deref for TlsStream<IO> {
+    type Target = tokio_rustls::server::TlsStream<IO>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl<T> DerefMut for TlsStream<T> {
+impl<IO> DerefMut for TlsStream<IO> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
 }
 
-impl<T: ActixStream> AsyncRead for TlsStream<T> {
+impl<IO: ActixStream> AsyncRead for TlsStream<IO> {
     fn poll_read(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -57,7 +58,7 @@ impl<T: ActixStream> AsyncRead for TlsStream<T> {
     }
 }
 
-impl<T: ActixStream> AsyncWrite for TlsStream<T> {
+impl<IO: ActixStream> AsyncWrite for TlsStream<IO> {
     fn poll_write(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -87,13 +88,13 @@ impl<T: ActixStream> AsyncWrite for TlsStream<T> {
     }
 }
 
-impl<T: ActixStream> ActixStream for TlsStream<T> {
+impl<IO: ActixStream> ActixStream for TlsStream<IO> {
     fn poll_read_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
-        T::poll_read_ready((&**self).get_ref().0, cx)
+        IO::poll_read_ready((&**self).get_ref().0, cx)
     }
 
     fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
-        T::poll_write_ready((&**self).get_ref().0, cx)
+        IO::poll_write_ready((&**self).get_ref().0, cx)
     }
 }
 
@@ -134,8 +135,8 @@ impl Clone for Acceptor {
     }
 }
 
-impl<T: ActixStream> ServiceFactory<T> for Acceptor {
-    type Response = TlsStream<T>;
+impl<IO: ActixStream> ServiceFactory<IO> for Acceptor {
+    type Response = TlsStream<IO>;
     type Error = TlsError<io::Error, Infallible>;
     type Config = ();
     type Service = AcceptorService;
@@ -155,17 +156,17 @@ impl<T: ActixStream> ServiceFactory<T> for Acceptor {
     }
 }
 
-/// Rustls based `Acceptor` service
+/// Rustls based acceptor service.
 pub struct AcceptorService {
     acceptor: TlsAcceptor,
     conns: Counter,
     handshake_timeout: Duration,
 }
 
-impl<T: ActixStream> Service<T> for AcceptorService {
-    type Response = TlsStream<T>;
+impl<IO: ActixStream> Service<IO> for AcceptorService {
+    type Response = TlsStream<IO>;
     type Error = TlsError<io::Error, Infallible>;
-    type Future = AcceptorServiceFut<T>;
+    type Future = AcceptFut<IO>;
 
     fn poll_ready(&self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         if self.conns.available(cx) {
@@ -175,8 +176,8 @@ impl<T: ActixStream> Service<T> for AcceptorService {
         }
     }
 
-    fn call(&self, req: T) -> Self::Future {
-        AcceptorServiceFut {
+    fn call(&self, req: IO) -> Self::Future {
+        AcceptFut {
             fut: self.acceptor.accept(req),
             timeout: sleep(self.handshake_timeout),
             _guard: self.conns.get(),
@@ -185,16 +186,18 @@ impl<T: ActixStream> Service<T> for AcceptorService {
 }
 
 pin_project! {
-    pub struct AcceptorServiceFut<T: ActixStream> {
-        fut: Accept<T>,
+    /// Accept future for Rustls service.
+    #[doc(hidden)]
+    pub struct AcceptFut<IO: ActixStream> {
+        fut: Accept<IO>,
         #[pin]
         timeout: Sleep,
         _guard: CounterGuard,
     }
 }
 
-impl<T: ActixStream> Future for AcceptorServiceFut<T> {
-    type Output = Result<TlsStream<T>, TlsError<io::Error, Infallible>>;
+impl<IO: ActixStream> Future for AcceptFut<IO> {
+    type Output = Result<TlsStream<IO>, TlsError<io::Error, Infallible>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let mut this = self.project();