+++ actix-tls/src/connect/tcp.rs
@@ -4,20 +4,20 @@ use actix_service::{Service, ServiceFactory};
 use super::{Address, Connect, ConnectError, ConnectServiceFactory, Connection, Resolver};
 
 /// Create TCP connector service.
-pub fn new_connector<T: Address + 'static>(
+pub fn new_connector<R: Address + 'static>(
     resolver: Resolver,
-) -> impl Service<Connect<T>, Response = Connection<T, TcpStream>, Error = ConnectError> + Clone
+) -> impl Service<Connect<R>, Response = Connection<R, TcpStream>, Error = ConnectError> + Clone
 {
     ConnectServiceFactory::new(resolver).service()
 }
 
 /// Create TCP connector service factory.
-pub fn new_connector_factory<T: Address + 'static>(
+pub fn new_connector_factory<R: Address + 'static>(
     resolver: Resolver,
 ) -> impl ServiceFactory<
-    Connect<T>,
+    Connect<R>,
     Config = (),
-    Response = Connection<T, TcpStream>,
+    Response = Connection<R, TcpStream>,
     Error = ConnectError,
     InitError = (),
 > + Clone {
@@ -25,17 +25,17 @@ pub fn new_connector_factory<T: Address + 'static>(
 }
 
 /// Create TCP connector service with default parameters.
-pub fn default_connector<T: Address + 'static>(
-) -> impl Service<Connect<T>, Response = Connection<T, TcpStream>, Error = ConnectError> + Clone
+pub fn default_connector<R: Address + 'static>(
+) -> impl Service<Connect<R>, Response = Connection<R, TcpStream>, Error = ConnectError> + Clone
 {
     new_connector(Resolver::Default)
 }
 
 /// Create TCP connector service factory with default parameters.
-pub fn default_connector_factory<T: Address + 'static>() -> impl ServiceFactory<
-    Connect<T>,
+pub fn default_connector_factory<R: Address + 'static>() -> impl ServiceFactory<
+    Connect<R>,
     Config = (),
-    Response = Connection<T, TcpStream>,
+    Response = Connection<R, TcpStream>,
     Error = ConnectError,
     InitError = (),
 > + Clone {